Restructured protocol and instance of it for data-flow with the model of TaskItem, logic inside AddTaskPresenter been refactored(if-else block inside textProcessing function realized via ternary operator)
