+++ ToDoList/ToDoList/Modules/AddTask/Core/Controllers/Protocol/AddTaskViewControllerInterface.swift
@@ -4,5 +4,7 @@
 //
 //  Created by Никита Данилович on 08.05.2023.
 //
-
-import Foundation
+protocol AddTaskViewControllerInterface: AnyObject {
+    var presenter: AddTaskPresenterInterface? { get }
+    func textFieldProcessed(with success: Bool)
+}
