+++ ToDoList/ToDoList/Modules/AddTask/Core/Presenter/AddTaskPresenter.swift
@@ -4,29 +4,23 @@
 //
 //  Created by Никита Данилович on 08.05.2023.
 //
-
-protocol AddTaskViewControllerInterface: AnyObject {
-    var presenter: AddTaskPresenterInterface? { get }
-    func textFieldProcessed(with success: Bool)
-}
-
-protocol AddTaskPresenterInterface: AnyObject {
-    func processTitleTextField(text: String)
-}
-
 class AddTaskPresenter: AddTaskPresenterInterface {
-   private unowned let view: AddTaskViewControllerInterface!
+    
+    private unowned let view: AddTaskViewControllerInterface
+    private let storage: StorageInterface
 
-    init(view: AddTaskViewControllerInterface) {
+    init(view: AddTaskViewControllerInterface, storage: StorageInterface) {
         self.view = view
+        self.storage = storage
     }
 
     func processTitleTextField(text: String) {
         let trimString = text.trimmingCharacters(in: .whitespaces)
-        if trimString != "" && trimString.count >= 2 {
-            view?.textFieldProcessed(with: true)
-        } else {
-            view?.textFieldProcessed(with: false)
-        }
+        let status = trimString != "" && trimString.count >= 2 ? true : false
+        view.textFieldProcessed(with: status)
+    }
+    
+    func addTaskToStorage(task: TaskItem) {
+        storage.addTask(status: task.status, title: task.title, description: task.description)
     }
 }
